+++ marvel-api-app/package.json
@@ -26,6 +26,7 @@
     "eslint": "^8.45.0",
     "eslint-plugin-react-hooks": "^4.6.0",
     "eslint-plugin-react-refresh": "^0.4.3",
+    "prettier": "^3.1.0",
     "typescript": "^5.0.2",
     "vite": "^4.4.5"
   }
