+++ marvel-api-app/src/App.tsx
@@ -1,13 +1,12 @@
-import CharacterForm from './components/CharacterForm/CharacterForm'
-import './App.css'
+import CharacterForm from './components/CharacterForm/CharacterForm';
+import './App.css';
 
 function App() {
-
   return (
     <>
       <CharacterForm></CharacterForm>
     </>
-  )
+  );
 }
 
-export default App
+export default App;
