+++ marvel-api-app/src/components/CharacterForm/CharacterForm.tsx
@@ -1,17 +1,17 @@
-import { useState } from "react";
-import { MD5 } from "crypto-js";
-import { ComicDataWrapper } from "../../marvel-api-types";
+import { useState } from 'react';
+import { MD5 } from 'crypto-js';
+import { ComicDataWrapper } from '../../marvel-api-types';
 
- 
 const CharacterForm = () => {
   const [searchParams, setSearchParams] = useState({
-    name: "",
-    comic: "",
+    name: '',
+    comic: '',
     // add more search parameters as needed
   });
 
-  const [responseData, setResponseData] = useState<ComicDataWrapper | null>(null);
-
+  const [responseData, setResponseData] = useState<ComicDataWrapper | null>(
+    null
+  );
 
   const handleInputChange = (event: React.ChangeEvent<HTMLInputElement>) => {
     const { name, value } = event.target;
@@ -24,7 +24,7 @@ const CharacterForm = () => {
   const handleSubmit = async (event: React.FormEvent<HTMLFormElement>) => {
     event.preventDefault();
 
-    const baseUrl = "https://gateway.marvel.com/v1/public/comics";
+    const baseUrl = 'https://gateway.marvel.com/v1/public/comics';
     const apiKey = import.meta.env.VITE_MARVEL_API_KEY;
     const privateKey = import.meta.env.VITE_MARVEL_PRIVATE_KEY;
 
@@ -39,12 +39,12 @@ const CharacterForm = () => {
 
     // Set headers
     const headers = {
-      Accept: "application/json",
+      Accept: 'application/json',
     };
 
     // Make the API call
     const response = await fetch(url, {
-      method: "GET",
+      method: 'GET',
       headers: headers,
     });
 
@@ -78,32 +78,34 @@ const CharacterForm = () => {
             </tr>
           </thead>
           <tbody>
-          {responseData?.data?.results?.map((result) => (
-  <tr key={result.id}>
-    <td>
-      {result.urls &&
-        result.urls
-          .filter((urlItem) => urlItem.type === "detail")
-          .map((detailLink) => (
-            <a key={detailLink.type} href={detailLink.url} target="_blank" rel="noopener noreferrer">
-              {result.title}
-            </a>
-          ))}
-    </td>
-    <td>{result.issueNumber}</td>
-    <td>
-      {result.thumbnail && (
-        <img
-          src={`${result.thumbnail.path}.${result.thumbnail.extension}`}
-          alt={result.title}
-          style={{ maxWidth: '100px' }}
-        />
-      )}
-    </td>
-  </tr>
-))}
-
-
+            {responseData?.data?.results?.map((result) => (
+              <tr key={result.id}>
+                <td>
+                  {result.urls &&
+                    result.urls
+                      .filter((urlItem) => urlItem.type === 'detail')
+                      .map((detailLink) => (
+                        <a
+                          key={detailLink.type}
+                          href={detailLink.url}
+                          target="_blank"
+                          rel="noopener noreferrer">
+                          {result.title}
+                        </a>
+                      ))}
+                </td>
+                <td>{result.issueNumber}</td>
+                <td>
+                  {result.thumbnail && (
+                    <img
+                      src={`${result.thumbnail.path}.${result.thumbnail.extension}`}
+                      alt={result.title}
+                      style={{ maxWidth: '100px' }}
+                    />
+                  )}
+                </td>
+              </tr>
+            ))}
           </tbody>
         </table>
       )}
